+++ docs/source/conf.py
@@ -1,7 +1,12 @@
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath("../.."))
+
 project = "TuViMCP"
 author = "Manh Ha Nguyen"
 copyright = "2026, nmhaaa3218"
-release = "0.4.0"
+release = "0.4.1"
 
 extensions = [
     "sphinx.ext.autodoc",